extras/requirements.txt
python-slugify==1.1.4
pony==0.6.1